refactor: document repository maturity plan and extract hosted agent domain
