+++ memory/repository-maturity.md
@@ -1,0 +1,28 @@
+# Repository Maturity
+
+## 本次发现的核心结构问题
+- `src/pages/chat.js` 与 `src/pages/assistant.js` 体量过大，页面职责与业务规则耦合严重。
+- `src/lib/tauri-api.js` 同时承担 transport、缓存、错误包装与部分页面依赖入口，边界不够清晰。
+- `scripts/dev-api.js` 过于庞大，路由、中间件、配置、命令逻辑混杂。
+- UI 层直接理解较多后端响应细节，adapter/view-model 边界不足。
+- chat/high-frequency 页面存在状态机复杂、时序敏感、体验问题容易反复出现的风险。
+
+## 已完成改造项
+- 新增 `src/lib/hosted-agent.js`，抽离 hosted Agent 常量、提示词、解析逻辑、动作文案。
+- `src/pages/chat.js` 改为消费 hosted-agent 模块，降低页面内联业务规则密度。
+- 删除 chat 自动滚动并禁用 chat 虚拟滚动，去掉隐式滚动接管链。
+- 统一 hosted 状态文案、系统提示持久化与状态展示反馈。
+
+## 后续建议
+- 继续拆 `src/pages/chat.js`：history/domain、hosted runtime/service、session event adapter。
+- 继续拆 `src/pages/assistant.js`：配置表单、状态机、工具调用、渲染层。
+- 为 `tauri-api.js` 增加 adapter/view-model 边界，减少页面直接消费 command 细节。
+- 分阶段拆 `scripts/dev-api.js`，优先 route dispatch 与 command handler 分离。
+- 补 lint / 类型检查或等价静态校验，治理 Vite import warning。
+
+## 约定的代码治理原则
+- 页面层只负责页面装配与交互编排。
+- 纯业务规则、提示词模板、解析逻辑优先抽到 `src/lib/`。
+- UI 不直接耦合底层后端 DTO，优先通过 adapter/view-model 消费。
+- 能删复杂逻辑就删，不为保留复杂实现而牺牲稳定性。
+- 巨型文件优先按“常量/纯函数/状态转换”顺序渐进拆分，避免一次性大重构。
